refactor: Merge appbar and body files into page
