--- packages/flutter_app/lib/pages/app_info_page/components/app_info_app_bar.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:themes/themes.dart';
-
-import '../../../gen/strings.g.dart';
-
-class AppInfoAppBar extends StatelessWidget implements PreferredSizeWidget {
-  const AppInfoAppBar({super.key});
-
-  @override
-  Size get preferredSize => appBarDefaultSize;
-
-  @override
-  Widget build(BuildContext context) {
-    final t = Translations.of(context);
-    return AppBar(
-      title: Text(t.appInfo.appBar.title),
-    );
-  }
-}
