--- packages/flutter_app/lib/pages/pinkie_mew/components/pinkie_mew_page_body.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-class PinkieMewPageBody extends HookConsumerWidget {
-  const PinkieMewPageBody({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Container();
-  }
-}
