+++ packages/themes/lib/src/altive_theme/altive_theme.dart
@@ -1,5 +1,4 @@
 export 'dark_theme_data.dart';
-export 'layout.dart';
 export 'light_theme_data.dart';
 export 'theme_extensions/theme_extensions.dart';
 export 'utils/build_context_extension.dart';
