--- packages/themes/lib/src/altive_theme/layout.dart
@@ -1,3 +1,0 @@
-import 'package:flutter/material.dart';
-
-const appBarDefaultSize = Size(double.infinity, kToolbarHeight);
